Fix Windows compile error, graph label visibility, mobile layout, detail panel width (#14, #9, #13, #12)
- Split POSIX-only daemon functions (Start/Stop/IsRunning) into
  daemon_posix.go with build tag, add Windows stub
- Lower graph label opacity threshold from 0.5 to 0.2 so most nodes
  are labeled; dim low-salience labels instead of hiding them
- Fix graph controls/legend overlap on mobile by repositioning and
  constraining widths
- Use clamp() for detail panel width instead of fixed 280px
