+++ internal/web/static/index.html
@@ -504,7 +504,7 @@
         .legend-line { width: 20px; height: 2px; flex-shrink: 0; border-radius: 1px; }
         .graph-detail {
             position: absolute; top: 20px; right: 20px;
-            width: 280px; max-height: calc(100% - 40px);
+            width: clamp(200px, 25%, 360px); max-height: calc(100% - 40px);
             background: var(--bg-secondary); border: 1px solid var(--border-color);
             border-radius: var(--radius-md); padding: 16px;
             z-index: 10; overflow-y: auto; display: none;
@@ -796,8 +796,11 @@
             .explore-view { padding: 12px 16px; }
             .explore-search { width: 140px; }
             .activity-drawer { width: 100vw; max-width: 100vw; }
-            .graph-controls { bottom: 10px; left: 10px; padding: 10px; }
-            .graph-legend { top: 10px; left: 10px; padding: 8px 10px; }
+            .graph-controls { bottom: 10px; left: 10px; right: 10px; padding: 10px; }
+            .graph-controls label { flex-wrap: wrap; }
+            .graph-controls input[type="range"] { width: 80px; }
+            .graph-legend { top: auto; bottom: 80px; left: 10px; padding: 8px 10px; max-width: calc(100vw - 20px); }
+            .graph-detail { width: calc(100vw - 20px); left: 10px; right: 10px; }
             .agent-grid { grid-template-columns: 1fr; }
             .agent-view { padding: 12px 16px; }
             .agent-header { flex-direction: column; align-items: flex-start; gap: 8px; }
@@ -1521,7 +1524,7 @@
             .text(function(d) { return (d.summary || '').slice(0, 30); })
             .attr('font-size', 10).attr('fill', '#cbd5e1')
             .attr('dx', function(d) { return Math.max(8, (d.salience || 0.5) * 40) + 4; }).attr('dy', 4)
-            .style('opacity', function(d) { return (d.salience || 0) > 0.5 ? 1 : 0; })
+            .style('opacity', function(d) { return (d.salience || 0) > 0.2 ? 0.9 : 0.4; })
             .style('pointer-events', 'none');
 
         var tooltip = document.getElementById('graphTooltip');
@@ -1550,7 +1553,7 @@
             tooltip.style.display = 'none';
             node.style('opacity', 1);
             link.style('opacity', function(d) { return Math.max(0.3, (d.strength || 0.5) * 0.8); });
-            label.style('opacity', function(d) { return (d.salience || 0) > 0.5 ? 1 : 0; });
+            label.style('opacity', function(d) { return (d.salience || 0) > 0.2 ? 0.9 : 0.4; });
         })
         .on('click', function(event, d) { event.stopPropagation(); showGraphDetail(d, validEdges, nodeMap); });
